Fix comparison of Short values with IN and CONTAINS in SQL
Resolves: #7578

+++ core/src/main/java/com/orientechnologies/orient/core/metadata/schema/OType.java
@@ -582,7 +582,7 @@ public enum OType {
 
   public static Number[] castComparableNumber(Number context, Number max) {
     // CHECK FOR CONVERSION
-    if (context instanceof Integer) {
+    if (context instanceof Short) {
       // SHORT
       if (max instanceof Integer)
         context = context.intValue();
